+++ src/js/common/utils/helpers.js
@@ -209,3 +209,21 @@ export function dateDiffDesc(date, userFromDate = null) {
 
   return 'less than a minute';
 }
+
+function isGaLoaded() {
+  return !!(window.ga && ga.create);
+}
+
+// google analytics client Id
+/* global gaClientId ga:true */
+export function gaClientId() {
+  let clientId;
+  if (isGaLoaded()) {
+    for (const data of ga.getAll()) {
+      if (data.get('cookieDomain') === 'vets.gov') {
+        clientId = data.get('clientId');
+      }
+    }
+  }
+  return clientId;
+}
